refactor(solana): one ExecutionDictionary behind a zama-host encode feature
host_mollusk.rs and the e2e live client each carried an identical interned-
dictionary builder for assembling FheExecuteArgs. It now lives in
zama-host::encode behind an 'encode' feature (mirroring 'decode': off-chain
only, never part of the SBF build), and both callers were compiled against
the feature before their copies were deleted.

Call sites hand the finished dictionary over through into_entries() instead
of reaching into the tuple field.

Part of fhevm-internal#1876 (M2).
