+++ solana/programs/zama-host/src/lib.rs
@@ -19,6 +19,8 @@ pub mod constants;
 /// Off-chain decoding of instruction data and event self-CPIs (feature `decode`).
 #[cfg(feature = "decode")]
 pub mod decode;
+#[cfg(feature = "encode")]
+pub mod encode;
 /// EIP-712 v4 verification of EVM-signed KMS / coprocessor certificates.
 pub mod eip712;
 /// Program-specific errors returned by ZamaHost instructions.
