+++ solana/scripts/e2e/live-client/Cargo.toml
@@ -21,7 +21,7 @@ solana-keypair = "3.1.2"
 ureq = { version = "2", features = ["json"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-zama-host = { path = "../../../programs/zama-host", features = ["no-entrypoint"] }
+zama-host = { path = "../../../programs/zama-host", features = ["no-entrypoint", "encode"] }
 zama-solana-acl = { path = "../../../crates/zama-solana-acl" }
 confidential-token = { path = "../../../programs/confidential-token", features = ["no-entrypoint"] }
